Add unit tests for dfsummary; fix bool-column quantile crash
Tests exercise summarize/compare/distance across numeric, categorical,
bool, datetime, missing/inf, MultiIndex, empty, and unhashable-index
DataFrames. Uncovered a real bug: bool columns are numeric per
is_numeric_dtype but pandas can't compute quantile() on bool Series,
which crashed column_stats (and anything comparing/diffing against a
DataFrame with a bool column). Fixed by casting the finite-value subset
to float before computing mean/median/std/quantiles.

+++ dfsummary.py
@@ -69,7 +69,7 @@ def _column_stats(s: pd.Series, quantiles=DEFAULT_QUANTILES) -> dict:
     stats["pct_inf"] = n_inf / n if n else np.nan
 
     if is_numeric:
-        finite = s[np.isfinite(s.astype(float))]
+        finite = s[np.isfinite(s.astype(float))].astype(float)  # bool dtype breaks quantile()
         n_zero = int((finite == 0).sum())
         n_negative = int((finite < 0).sum())
         n_positive = int((finite > 0).sum())
